feat: added separate component for popup, resolved req changes
