+++ src/components/InvitePopup/InvitePopup.scss
@@ -1,0 +1,75 @@
+.modal-container {
+  width: 500px;
+  font-size: 20px;
+  background-color: #fafafa;
+  border-radius: 8px;
+  box-shadow: 2px 2px 8px rgba(0, 0, 0, 0.5);
+  .success-container {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    padding: 0px 50px 40px 50px;
+    backdrop-filter: blur(4px);
+    .content {
+      display: flex;
+      flex-direction: column;
+      width: -webkit-fill-available;
+    }
+    .header-container {
+      font-weight: 700;
+      font-size: xx-large;
+      margin-bottom: 20px;
+    }
+    .text-description {
+      text-align: center;
+      font-size: 16px;
+      margin-top: 10px;
+      opacity: 0.8;
+    }
+  }
+  .failure-container {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    padding: 0px 50px 40px 50px;
+    backdrop-filter: blur(4px);
+    .content {
+      display: flex;
+      flex-direction: column;
+      width: -webkit-fill-available;
+    }
+    .header-container {
+      font-weight: 700;
+      font-size: xx-large;
+      margin-bottom: 20px;
+    }
+    .text-description {
+      text-align: center;
+      font-size: medium;
+      margin-top: 10px;
+    }
+  }
+  .close-btn-container {
+    display: flex;
+    justify-content: flex-end;
+    padding: 10px 10px 0px 0px;
+    .close-modal {
+      color: var(--grey3Color);
+      padding: 0;
+      border: none;
+      background: none;
+      .close-icon {
+        height: 20px;
+        width: 20px;
+      }
+      &:hover {
+        cursor: pointer;
+        color: var(--grey5Color);
+      }
+    }
+  }
+}
+
+.invite-popup-overlay {
+  background: rgba(0, 0, 0, 0.5);
+}
