+++ src/components/InvitePopup/index.ts
@@ -1,0 +1,3 @@
+import InvitePopup from "./InvitePopup";
+
+export default InvitePopup;
