+++ src/components/InvitePopup/model.ts
@@ -1,0 +1,5 @@
+export default interface IInviteProps {
+  isOpen: boolean;
+  memberLoading: boolean;
+  isData: boolean;
+}
